Update changelog and working demos; get rid of some warnings; update copyright notice everywhere

+++ kkapturedll/avi_videoencoder_dshow.cpp
@@ -1,5 +1,5 @@
 /* kkapture: intrusive demo video capturing.
- * Copyright (c) 2005-2006 Fabian "ryg/farbrausch" Giesen.
+ * Copyright (c) 2005-2009 Fabian "ryg/farbrausch" Giesen.
  *
  * This program is free software; you can redistribute and/or modify it under
  * the terms of the Artistic License, Version 2.0beta5, or (at your opinion)
@@ -198,7 +198,7 @@ public:
     VIDEOINFOHEADER *vi = (VIDEOINFOHEADER *) m_mt.Format();
     sample->GetPointer(&dataPtr);
 
-    memcpy(dataPtr,sourceData,min(sample->GetSize(),vi->bmiHeader.biSizeImage));
+    memcpy(dataPtr,sourceData,min(sample->GetSize(),(int) vi->bmiHeader.biSizeImage));
 
     // set time and sync point
     REFERENCE_TIME start,end;
