+++ kkapturedll/avi_videoencoder_vfw.cpp
@@ -1,5 +1,5 @@
 /* kkapture: intrusive demo video capturing.
- * Copyright (c) 2005-2006 Fabian "ryg/farbrausch" Giesen.
+ * Copyright (c) 2005-2009 Fabian "ryg/farbrausch" Giesen.
  *
  * This program is free software; you can redistribute and/or modify it under
  * the terms of the Artistic License, Version 2.0beta5, or (at your opinion)
@@ -275,7 +275,7 @@ AVIVideoEncoderVFW::AVIVideoEncoderVFW(const char *name,int _fpsNum,int _fpsDeno
   
   // determine file base name
   strcpy_s(d->basename,sizeof(d->basename),name);
-  for(int i=strlen(d->basename)-1;i>=0;i--)
+  for(int i=(int) strlen(d->basename)-1;i>=0;i--)
   {
     if(d->basename[i] == '/' || d->basename[i] == '\\')
       break;
